[Cleanup] Eliminate warnings (#393)
This is a combined patch for eliminating most of the warning messages.

+++ src/dft/mkunroll.c
@@ -11,8 +11,8 @@
 #define CONFIGMAX 4
 
 char *replaceAll(const char *in, const char *pat, const char *replace) {
-  const int replaceLen = strlen(replace);
-  const int patLen = strlen(pat);
+  const int replaceLen = (int)strlen(replace);
+  const int patLen = (int)strlen(pat);
 
   char *str = malloc(strlen(in)+1);
   strcpy(str, in);
@@ -21,8 +21,8 @@ char *replaceAll(const char *in, const char *pat, const char *replace) {
     char *p = strstr(str, pat);
     if (p == NULL) return str;
 
-    int replace_pos = p - str;
-    int tail_len = strlen(p + patLen);
+    int replace_pos = (int)(p - str);
+    int tail_len = (int)strlen(p + patLen);
 
     char *newstr = malloc(strlen(str) + (replaceLen - patLen) + 1);
